+++ src/main/scala/com/raphtory/core/actors/orchestration/componentconnector/PartitionConnector.scala
@@ -6,8 +6,7 @@ import akka.cluster.pubsub.DistributedPubSubMediator
 import akka.pattern.ask
 import com.raphtory.core.actors.RaphtoryActor.{partitionMachineCount, partitionsPerMachine}
 import com.raphtory.core.actors.orchestration.clustermanager.WatchDog.Message.RequestPartitionId
-import com.raphtory.core.actors.partitionmanager.workers.{IngestionWorker, ReaderWorker}
-import com.raphtory.core.actors.partitionmanager.PartitionManager
+import com.raphtory.core.actors.partitionmanager.{IngestionWorker, PartitionManager, ReaderWorker}
 import com.raphtory.core.model.graph.GraphPartition
 import com.raphtory.core.model.implementations.objectgraph.ObjectBasedPartition
