+++ src/main/scala/com/raphtory/core/model/implementations/objectgraph/entities/internal/ImmutableProperty.scala
@@ -1,6 +1,6 @@
 package com.raphtory.core.model.implementations.objectgraph.entities.internal
 
-import com.raphtory.core.actors.partitionmanager.workers.ParquetProperty
+import com.raphtory.core.actors.partitionmanager.ParquetProperty
 
 class ImmutableProperty(creationTime: Long, value: Any) extends Property {
   var earliestTime: Long                = creationTime
