+++ src/main/scala/com/raphtory/core/model/implementations/objectgraph/entities/internal/RaphtoryEdge.scala
@@ -1,6 +1,6 @@
 package com.raphtory.core.model.implementations.objectgraph.entities.internal
 
-import com.raphtory.core.actors.partitionmanager.workers.ParquetEdge
+import com.raphtory.core.actors.partitionmanager.ParquetEdge
 import com.raphtory.core.model.graph.GraphPartition
 
 import scala.collection.mutable
